requirements.txt
fastapi
uvicorn
openai
langchain>=0.1.14
langchain-community>=0.0.11
langchain-openai>=0.0.6
tiktoken==0.9.0
faiss-cpu
pymysql
pydantic
python-dotenv
